+++ src/cli.ts
@@ -19,6 +19,7 @@ import { blacklistCommand } from './commands/blacklist'
 import { checkCommand } from './commands/check'
 import { diffCommand } from './commands/diff'
 import { statusCommand } from './commands/status'
+import { inspectCommand } from './commands/inspect'
 import { doctorCommand } from './commands/doctor'
 import { completionCommand } from './commands/completion'
 import { upgradeCommand, formatUpdateHint, formatSkillUpdateReminder } from './commands/upgrade'
@@ -258,6 +259,7 @@ program.addCommand(diffCommand)
 
 // Register status command
 program.addCommand(statusCommand)
+program.addCommand(inspectCommand)
 program.addCommand(doctorCommand)
 program.addCommand(completionCommand)
 program.addCommand(upgradeCommand)